pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mixedbread-ai-langchain"
version = "0.1.0"
description = "Mixedbread AI integration for LangChain, providing embedding, document loading, and reranking components."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
keywords = ["langchain", "mixedbread-ai", "embedding", "reranking", "document-loader", "nlp", "llm"]
authors = [{ name = "Mixedbread AI", email = "support@mixedbread.com" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "langchain-core>=0.1.0",
    "mixedbread-ai>=0.3.1",
    "httpx>=0.24.0"
]

[project.urls]
Homepage = "https://github.com/mixedbread-ai/mixedbread-ai-langchain"
Documentation = "https://github.com/mixedbread-ai/mixedbread-ai-langchain#readme"
Issues = "https://github.com/mixedbread-ai/mixedbread-ai-langchain/issues"
Source = "https://github.com/mixedbread-ai/mixedbread-ai-langchain"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "tqdm",  # For progress bars in embedders
]

[tool.pytest.ini_options]
markers = [
    "integration: integration tests that call the Mixedbread AI API (deselect with '-m \"not integration\"')",
]
log_cli = true

